- Warning fix.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@29371 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/engine/3d/collection.h
@@ -83,7 +83,7 @@ public:
       return;
 
     csWeakRefArray<iObject> copy;
-    for(int i=0; i<Children->GetSize(); i++)
+    for(size_t i=0; i<Children->GetSize(); i++)
     {
       copy.Push(Children->Get(i));
     }
